Añadir ejemplos al yaml.

+++ src/main/java/com/svalero/vuelosapi/dto/AirplanePatchDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AirplanePatchDto {
     private String field;
-    private int maxSpeed;
+    private float maxSpeed;
 }
